requirements.txt
streamlit==1.35.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.4.2
ollama==0.5.1
tqdm==4.66.4
transformers==4.41.1
torch==2.3.0
